added best selection and errol4
This commit combines two large changes: 1) all versions of Errol have been
modified to try to provide the "best" and optimal output and 2) the next
version of the algorithm -- Errol4 -- more than doubles the performance
compared to Errol3 (bringing its performance more in line with Grisu3). The
changes are explained in further detail below.

Best Selection. The previous versions of Errol selected any available optimal
(or shortest) output that rounds to the same floating point number. Now, the
algorithm selects the optimal number that falls closest to the floating point
number. In the event that two optimal values are equal distant from the
floating point number, the even value is chosen (round to even). The test code
was updated accordingly to verify that the output produced by Errol exactly
matches Dragon4.

IMPORTANT: Errol provides a best effort to choose the best output but does not
guarantee a correct selection. When testing with Errol3 and Errol4, we tested
one billion random inputs and found no failures. Therefore, we only claim to
select the best number with extremely high probability. Even if the best
selection fails, we always select a correct output (falls in the rounding
interval). We are currently working on an enumation based technique to
guarantee best output.

Errol4. The latest version of Errol4 provides a 2x performance improvement over
Errol3. Taking into account the corrected benchmarks (where Errol3 is about
2.4x slower than Grisu3), Errol4 averages about 10% slower than Grisu3 when
tested on a Intel Core i7-3367U.